Move ProgressBarListener to cli.internal package
